refactor anisette manager to own file
